Renamed csv_to_AstropyTable.py to adhere to pep8 standards for module names
